Simplify/improve animation of remaining words on successful guess (#3)
This minimizes and improves the animated "movement" that happens when
getting a successful guess. Previously all items would move. With this
change the remaining unguessed items on the top row move into the spots
made vacant by guessed items. This is what happens on the NYT game.

Note that the NYT game also animates the guess items up towards the top,
which this does not do.